Restructured project a bit
Pulled apart fractal rendering code and coloring algorithms to make my life less of a hell...
