+++ src/go.mod
@@ -1,0 +1,3 @@
+module utils
+
+go 1.17
